Move filter & blend aliases to arcade.gl.types
* Use PyGLenum per @cspotcode's suggestion

* Remove use of literals

* Move to arcade.gl.types

* Import TypeAlias from typing_extensions since it's from 3.10
