+++ arcade/sprite_list/sprite_list.py
@@ -31,7 +31,8 @@ from arcade import (
     get_window,
     gl,
 )
-from arcade.types import Color, RGBA255, OpenGlFilter, BlendFunction
+from arcade.types import Color, RGBA255
+from arcade.gl.types import OpenGlFilter, BlendFunction
 from arcade.gl.buffer import Buffer
 from arcade.gl.vertex_array import Geometry
